+++ scripts/macros/2024/spells/callLightning.js
@@ -1,0 +1,7 @@
+import {callLightning as callLightningLegacy} from '../../spells/callLightning.js';
+export let callLightning = {
+    name: 'Call Lightning',
+    version: '1.1.19',
+    rules: 'modern',
+    midi: callLightningLegacy.midi
+};