+++ scripts/macros/2024/spells/chainLightning.js
@@ -1,0 +1,9 @@
+import {chainLightning as chainLightningLegacy} from '../../spells/chainLightning.js';
+export let chainLightning = {
+    name: 'Chain Lightning',
+    version: '1.1.19',
+    rules: 'modern',
+    midi: chainLightningLegacy.midi,
+    config: chainLightningLegacy.config,
+    hasAnimation: true
+};